Improve: make the pattern searching match the way of pyinstaller.
The searching algorithm used now is fetch the magic pattern by fixed offset from the end of file. But according the source code of pyinstaller it should be searching from end to front, and I found some sample which added extra overlay to CArchive and stil works. So I rewrite some code to match the original algorithm.